moved tests to actual location in existing test files; added mock connection fixture
