+++ version.py
@@ -1,3 +1,3 @@
 # Update this for the versions
 # Don't change the forth version number from None
-VERSION = (1, 2, 4, 37)
+VERSION = (1, 2, 4, 39)
